Edit / Save / Delete

+++ ui/src/app/table/table.component.ts
@@ -22,26 +22,37 @@ export class TableComponent {
   vendordata: Vendor[];
   vendorcolumns: string[] = ["vendorId", "firmName", "gstnumber"];
   vendorUrl: string = "resource/vendors";
+  selected: Vendor;
+  vendorEmpty: Vendor = new Vendor(0,'','');
+//  showTable: boolean = false;
 
   constructor(private app: AppService,
               private atService: AdventureTimeService,
               private vendorService: VendorService,
               private http: HttpClient,
               private router: Router) {
-     this.columns = this.vendorService.getColumns();
-     this.characters = this.atService.getCharacters();
+     console.log("Inside vendor constructor")
+     this.getAllVendors();
+     this.selected = this.vendorEmpty;
   }
 
   authenticated() {
-     if(this.app.authenticated) {
-        this.getAllVendors();
-     }
+     console.log("Inside authenticated");
     // return true;
      return this.app.authenticated;
   }
 
   getAllVendors() {
-     this.vendordata = this.vendorService.getAllVendors();
+//     this.showTable = false;
+     this.vendorService.getAllVendors().subscribe(res => this.vendordata = res as Vendor[]);
+//     setTimeout(()=>{this.showTable = true}, 0);
+     //var vendor1 = new Vendor(1, 'MyFirm', 'MYGST');
+     //var vendor2 = new Vendor(2, 'MyFirm', 'MYGST');
+     //var vendorArray = new Array<Vendor>();
+     //vendorArray[0] = vendor1;
+     //vendorArray[1] = vendor2;
+     //this.vendordata = vendorArray;
+     this.selected = this.vendorEmpty;
   }
 
   navigate(url: string) {
@@ -58,12 +69,27 @@ export class TableComponent {
   }
 
   deleteVendor(id: number, idx: number) {
-    this.vendorService.deleteVendor(id);
-    this.vendordata.splice(idx, 1);
+    if(confirm("Are you sure to delete vendor with id : "+id)) {
+      console.log("Deleting the vendor");
+      this.vendorService.deleteVendor(id);
+      this.vendordata.splice(idx, 1);
+    }
   }
 
   updateVendor(id: number, vendor: Vendor) {
+    this.selected = vendor;
+  }
+
+  saveVendor(id: number, vendor: Vendor) {
     this.vendorService.updateVendor(id, vendor);
+    this.selected = this.vendorEmpty;
+  }
+
+  isselected(vendor: Vendor) {
+    if(this.selected.vendorId == vendor.vendorId) {
+      return true;
+    }
+    return false;
   }
 
   private handleError (error: Response | any) {
